The big refactoring and featur-add-ing of the enhanced textarea.
Refactored the key tracking code to look for a set of keys, then preventDefault, then get either two string (both sides of the cursor) or three strings (left of the selection, the selection, and right of the selection), and then finally have different actions for different keys. You can now tab, command + ], command + [, and when you hit return the next line should be at the same indention point as the one before it.